Improve 'move thread' post cleanup logic
This change removes unnecessary complexity from the post cleanup
logic in the 'move thread' command. Only the root post in the
thread is deleted via the API now. The replies to that thread are
automatically marked as deleted from this process.

This removes complexity and lowers the possibility of an error
being encountered.