Refactor post sorting logic and add metadata
This cleans up the post sorting logic required before moving a
thread with Wrangler. New post list metadata is now built on the
fly based on the contents of the original thread. This may be used
later for displaying thread information or for locking down plugin
usage based on thread context.

Lastly, the thread-move summary message now contains a snippet of
the content of the original thread root message for more clarity
on what was moved.