Preserve message timestamp when attaching it to a thread (#55)
This changes the behavior of `/wrangler attach message` to preserve
the timestamp of the original message.

The timestamp was originally ignored and the new message showed up
in the thread based on the time when the attach command was run.
This was done as new messages show up at the bottom of channels
even when they are created with a timestamp sometime before previous
messages in the thread. As this is the current Mattermost behavior,
it initially felt less confusing to proceed in this manner. Upon
further testing, preserving the timestamp of the original message
is far more useful even when dealing with the less-than-ideal
channel behavior mentioned above.