Ensure root posts are in redux
Some scenarios where the post dropdown could be used in the
Mattermost webapp could result in a state where the post of a
thread was available in redux, but the root post was not available.
Some of these cases were not properly handled in the plugin.

This change adds logic for ensuring root posts and their threads
are pulled into redux so that the Wrangler webapp can properly
display and manage them.