Add webapp option to attach messages to threads
This introduces the ability to attach unthreaded messages to
threads via the webapp UI. This is done by introducing a new
asynchronous flow where the user selects a message to attach and
then later attaches it to a selected thread by using post dropdowns.

While this flow is underway, a new sidebar element will remind
users of the context of their in-progress message attachment and
will allow them to cancel it if they wish.