Add webapp option to target channel to copy messages to
This introduces the ability to easily copy multiple messages to a
targeted channel via the webapp UI. This is done by introducing a
new asynchronous flow where the user selects a channel to target
and then uses post dropdowns to copy messages and threads.

While this flow is underway, a new sidebar element will remind
users of the context of their in-progress message copying and
will allow them to cancel it when they are done.