Add `merge thread` command (#162)
This new optional command allows for merging the messages of thread
into another existing thread. The message timestamps are preserved
so the resulting thread will contain all messages from both threads
in the order in which they were originally created.

Due to how this command functions, it is turned off by default and
must be enabled in the plugin configuration before it can be used.
The existing plugin permissions for moving threads also apply to
merging threads if the command is enabled.