Disable modal button when processing (#184)
When the Wrangler modal is being used to move or copy messages
there can be a processing delay between when the action button
is clicked and when the modal disappears. This change updates the
styling and behavior of the action button so that it becomes
disabled while processing occurs and the button text now indicates
that work is being done by the plugin.