Check for undefined teams in webapp (#185)
When building certain webapp components a list of teams the user
belongs to is required. When looking up these teams there appears
to be certain edge cases where a team object may not be present
in redux. It's unclear if this should be happening, but this change
checks if the team is undefined to prevent plugin crashes when a
team is missing.